backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
ruff==0.1.6
sqlmodel==0.0.14
httpx==0.25.0
